nullable silo data for eth silo
