+++ src/repository/postgres/models/index.js
@@ -1,12 +1,13 @@
 'use strict';
 
-require('dotenv').config();
+const EnvUtil = require('../../../utils/env');
 const fs = require('fs');
 const path = require('path');
 const Sequelize = require('sequelize');
 const process = require('process');
+
 const basename = path.basename(__filename);
-const env = process.env.NODE_ENV || 'development';
+const env = EnvUtil.getDeploymentEnv() || 'local';
 const config = require(__dirname + '/../config/config.js')[env];
 const db = {};
