+++ src/repository/postgres/seeders/20240725030936-silo-tokens.js
@@ -5,16 +5,16 @@ const db = require('../models');
 const ContractGetters = require('../../../datasources/contracts/contract-getters');
 const { C } = require('../../../constants/runtime-constants');
 const AlchemyUtil = require('../../../datasources/alchemy');
+const PromiseUtil = require('../../../utils/promise');
 
 const c = C('eth');
+const tokens = [c.BEAN, c.BEANWETH, c.BEANWSTETH, c.BEAN3CRV, c.UNRIPE_BEAN, c.UNRIPE_LP];
 
 /** @type {import('sequelize-cli').Migration} */
 module.exports = {
   async up(queryInterface, Sequelize) {
     await AlchemyUtil.ready(c.CHAIN);
     const beanstalk = await ContractGetters.getBeanstalk(c);
-
-    const tokens = [c.BEAN, c.BEANWETH, c.BEANWSTETH, c.BEAN3CRV, c.UNRIPE_BEAN, c.UNRIPE_LP];
 
     // Gets tokens that have already been populated
     const existingTokens = await db.sequelize.models.Token.findAll({
@@ -32,12 +32,14 @@ module.exports = {
       const rows = [];
       for (const token of newTokens) {
         const erc20 = await get(token);
-        // TODO: if any of these revert, as will be the case after migration, they should be set null in the table.
-        const [name, symbol, decimals, stalkEarnedPerSeason, stemTip, totalDeposited, totalDepositedBdv] =
-          await Promise.all([
-            erc20.name(),
-            erc20.symbol(),
-            (async () => Number(await erc20.decimals()))(),
+        const [name, symbol, decimals] = await Promise.all([
+          erc20.name(),
+          erc20.symbol(),
+          (async () => Number(await erc20.decimals()))()
+        ]);
+        const [bdv, stalkEarnedPerSeason, stemTip, totalDeposited, totalDepositedBdv] = await Promise.all(
+          [
+            (async () => BigInt(await beanstalk.bdv(token, BigInt(10 ** decimals))))(),
             (async () => {
               const tokenSettings = await beanstalk.tokenSettings(token);
               return BigInt(tokenSettings.stalkEarnedPerSeason);
@@ -45,8 +47,9 @@ module.exports = {
             (async () => BigInt(await beanstalk.stemTipForToken(token)))(),
             (async () => BigInt(await beanstalk.getTotalDeposited(token)))(),
             (async () => BigInt(await beanstalk.getTotalDepositedBdv(token)))()
-          ]);
-        const bdv = BigInt(await beanstalk.bdv(token, BigInt(10 ** decimals)));
+            // If any revert, they return null instead
+          ].map(PromiseUtil.nullOnReject)
+        );
         rows.push({
           address: token,
           name,
@@ -69,7 +72,7 @@ module.exports = {
 
   async down(queryInterface, Sequelize) {
     await queryInterface.bulkDelete('token', {
-      token: {
+      address: {
         [Sequelize.Op.in]: tokens
       }
     });
